WIP remove CarValidationService and its interface
