--- CarRental/CarRental.Logic/Services/IServices/ICarValidationService.cs
@@ -1,8 +1,0 @@
-﻿using CarRental.Logic.Models;
-
-namespace CarRental.Logic.Services.IServices;
-public interface ICarValidationService
-{
-    bool IsValidForCreate(CarViewModel car);
-    bool IsAllValid(CarViewModel car);
-}
